chore: renamed index to token
